Renamed module 'os_traits' -> 'unix_string'
