+++ backhand/src/entry.rs
@@ -1,7 +1,7 @@
 use std::ffi::OsStr;
 use std::fmt;
 
-use crate::os_traits::OsStrExt;
+use crate::unix_string::OsStrExt;
 use crate::data::Added;
 use crate::dir::{Dir, DirEntry};
 use crate::inode::{
