+++ backhand/src/lib.rs
@@ -75,7 +75,7 @@ mod kinds;
 mod metadata;
 mod reader;
 mod squashfs;
-mod os_traits;
+mod unix_string;
 
 pub use crate::data::DataSize;
 pub use crate::error::BackhandError;
