+++ backhand/src/unix_string.rs
